separate nft winners selection
