+++ launchpad-with-nft/src/nft_winners_selection.rs
@@ -1,9 +1,8 @@
 elrond_wasm::imports!();
 elrond_wasm::derive_imports!();
 
-use elrond_wasm::elrond_codec::TopEncode;
 use launchpad_common::{
-    ongoing_operation::{OngoingOperationType, CONTINUE_OP, STOP_OP},
+    ongoing_operation::{CONTINUE_OP, STOP_OP},
     random::Random,
 };
 
@@ -20,23 +19,7 @@ pub trait NftWinnersSelectionModule:
     + crate::confirm_nft::ConfirmNftModule
     + crate::mystery_sft::MysterySftModule
 {
-    #[endpoint(selectNftWinners)]
-    fn select_nft_winners(&self) -> OperationCompletionStatus {
-        self.require_winner_selection_period();
-
-        let flags_mapper = self.flags();
-        let mut flags = flags_mapper.get();
-        require!(
-            flags.were_winners_selected,
-            "Must select winners for base launchpad first"
-        );
-        require!(
-            !flags.was_additional_step_completed,
-            "Already selected NFT winners"
-        );
-
-        let mut rng: Random<Self::Api> = self.load_additional_selection_operation();
-
+    fn select_nft_winners(&self, rng: &mut Random<Self::Api>) -> OperationCompletionStatus {
         let mut all_users_mapper = self.confirmed_nft_user_list();
         let mut nft_winners_mapper = self.nft_selection_winners();
 
@@ -44,7 +27,7 @@ pub trait NftWinnersSelectionModule:
         let mut winners_selected = nft_winners_mapper.len();
         let total_available_nfts = self.total_available_nfts().get();
 
-        let run_result = self.run_while_it_has_gas(|| {
+        self.run_while_it_has_gas(|| {
             if users_left == 0 || winners_selected == total_available_nfts {
                 return STOP_OP;
             }
@@ -59,28 +42,18 @@ pub trait NftWinnersSelectionModule:
             winners_selected += 1;
 
             CONTINUE_OP
-        });
+        })
+    }
 
-        match run_result {
-            OperationCompletionStatus::InterruptedBeforeOutOfGas => {
-                let mut encoded_rng = ManagedBuffer::new();
-                let _ = rng.top_encode(&mut encoded_rng);
+    #[view(hasUserConfirmedNft)]
+    fn has_user_confirmed_nft(&self, user: ManagedAddress) -> bool {
+        self.confirmed_nft_user_list().contains(&user)
+            || self.nft_selection_winners().contains(&user)
+    }
 
-                self.save_progress(&OngoingOperationType::AdditionalSelection {
-                    encoded_data: encoded_rng,
-                });
-            }
-            OperationCompletionStatus::Completed => {
-                flags.was_additional_step_completed = true;
-                flags_mapper.set(&flags);
-
-                let nft_cost = self.nft_cost().get();
-                let claimable_nft_payment = nft_cost.amount * winners_selected as u32;
-                self.claimable_nft_payment().set(&claimable_nft_payment);
-            }
-        }
-
-        run_result
+    #[view(hasUserWonNft)]
+    fn has_user_won_nft(&self, user: ManagedAddress) -> bool {
+        self.nft_selection_winners().contains(&user)
     }
 
     #[storage_mapper("nftSelectionWinners")]
